+++ pkg/common/framework/modifiers/doc.go
@@ -1,2 +1,2 @@
-// Package modifiers ...
+// Package modifiers contains plan modifier related functions to be used with resource and data sources written using Terraform Provider Framework
 package modifiers