+++ pkg/providers/cloudwrapper/provider.go
@@ -24,6 +24,7 @@ func NewSubprovider() *Subprovider {
 // Resources returns terraform resources for cloudwrapper
 func (p *Subprovider) Resources() []func() resource.Resource {
 	return []func() resource.Resource{
+		NewActivationResource,
 		NewConfigurationResource,
 	}
 }