+++ pkg/providers/cloudwrapper/testdata/TestResActivation/create_missing_required.tf
@@ -1,0 +1,3 @@
+resource "akamai_cloudwrapper_activation" "act" {
+  config_id = 123
+}