+++ pkg/providers/cloudwrapper/testdata/TestResActivation/update_timeout.tf
@@ -1,0 +1,8 @@
+resource "akamai_cloudwrapper_activation" "act" {
+  config_id = 123
+  revision  = "8b92934d68d69621153c"
+  timeouts {
+    create = "2s"
+    update = "1s"
+  }
+}